# Reduce the allocations of gausslegendre for moderate n and add a flag for Aqua