--- a/src/gausslegendre.jl
+++ b/src/gausslegendre.jl
@@ -262,7 +262,7 @@ function leg_initial_guess(n)
     # Returns an approximation of the first n÷2+1 roots of the Legendre polynomial.
     #  The following is equivalent to "x0=asy(n);x = x0[1:n ÷ 2 + 1]" but it avoids unnecessary calculations.
 
-    m = (n + 1) >> 1
+    m = (n ÷2) +1
     a = besselZeroRoots(m)
     rmul!(a, 1 / (n + 0.5))
     x = legpts_nodes(n, a)