--- a/src/gausslegendre.jl
+++ b/src/gausslegendre.jl
@@ -258,7 +258,7 @@ end
     end
 end
 
-function leg_initial_guess(n)
+@inline function leg_initial_guess(n)
     # Returns an approximation of the first n÷2+1 roots of the Legendre polynomial.
     #  The following is equivalent to "x0=asy(n);x = x0[1:n ÷ 2 + 1]" but it avoids unnecessary calculations.